Add LanguageServerManager.from_languages (moved from Project)
